Data.TimeChunkBuilder

Description

Monad for the safe and efficient construction of continuous chunks of time values.

Chunk builder

data TimeChunkBuilder e a

Time Chunk builder monad. See the builder operations below to see what you can do in this Monad.

buildChunk

Arguments

:: Int
Initial time

-> Int
Chunk size

-> Int
Time margin

-> TimeChunkBuilder e ()
Builder operations

-> IO (TimeChunkState e, Either (TimeChunkEnv e) (Vector (Maybe e)))

Build a time chunk using the TimeChunkBuilder monad. The chunk is returned as a vector with potentially empty values (Nothing). If the chunk building is unfinished, it returns an environment that you can use to resume the chunk building using buildChunkFromEnv. In addition, values inserted with the chunk builder with time in the future or the past are returned in a TimeChunkState value.

data TimeChunkState e

Collection of values from the past and future.

Constructors

TCS

Fields

valuesFromThePast :: [(Int, e)]

valuesFromTheFuture :: [(Int, e)]

Resuming buildings

data TimeChunkEnv e

A Time Chunk Builder environment. You can use a value of this type to resume building a chunk with buildChunkFromEnv. The type is annotated with the type of the values in the chunk.

buildChunkFromEnv :: TimeChunkEnv e -> TimeChunkBuilder e () -> IO (TimeChunkState e, Maybe (Vector (Maybe e)))

Resume a chunk building using a supplied environment. As in buildChunk, the building might not finish. If this is the case, it returns Nothing instead of the vector. The chunk building can be continued using the same environment again.

Builder operations

insertValue

Arguments

:: Int
Time

-> e
Value

-> TimeChunkBuilder e ()

Insert a value in the chunk. If the value is outside of the time margin (i.e. later than the chunk final time plus the time margin parameter provided in buildChunk), the chunk building will be automatically finished.

finishChunkBuilding :: TimeChunkBuilder e ()

Finish the chunk building.